fix: 서비스 중복 내용 삭제 1
