+++ src/main/java/com/example/swip/service/impl/ChatServerServiceImpl.java
@@ -23,6 +23,7 @@ import com.mysema.commons.lang.Pair;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -41,8 +42,8 @@ public class ChatServerServiceImpl implements ChatServerService {
 
     private final StudyRepository studyRepository;
     private final UserRepository userRepository;
-    private final UserService userService;
     private final UserStudyRepository userStudyRepository;
+
     @Value("${swyp.chat.server.uri}")
     private String reqUserURL;
     public Pair<String, Integer> postUser(ChatProfileRequest chatProfileRequest){
@@ -432,14 +433,29 @@ public class ChatServerServiceImpl implements ChatServerService {
 
     private Pair<String, Integer> setUserStatusAndReturnPair(Long userId, Integer status_num, ChatStatus chatStatus) {
 
-        User user = userService.findUserById(userId);
+        User user = userRepository.findById(userId).orElse(null);
         if(chatStatus==ChatStatus.Need_create || chatStatus==ChatStatus.Need_update) {
-            userService.setChatStatus(user, status_num, chatStatus);
+            setChatStatus(user, status_num, chatStatus);
             return Pair.of("Failed to sync user data", 500);
         } else{
-            userService.setChatStatus(user, status_num, chatStatus);
+            setChatStatus(user, status_num, chatStatus);
             return Pair.of("Complete to sync user data", 200);
         }
+    }
+
+    @Transactional
+    public void setChatStatus(Object obj, Integer status_num, ChatStatus defaultStatus) {
+        if (status_num == 200)
+            setUserOrStudyChatStatus(obj, ChatStatus.Clear);
+        else
+            setUserOrStudyChatStatus(obj, defaultStatus);
+    }
+
+    private void setUserOrStudyChatStatus(Object obj, ChatStatus status) {
+        if(obj instanceof User)
+            ((User) obj).setChat_status(status);
+        else if (obj instanceof Study)
+            ((Study) obj).setChat_status(status);
     }
 
     private Pair<String, Integer> setStudyStatusAndReturnPair(Long studyId, Integer status_num, ChatStatus chatStatus) {
